linux-server/requirements.txt
fastapi
uvicorn
python-socketio
pynput
mss
Pillow
PyGObject
python-uinput
dbus-python
jeepney